tests/wpt/tests/permissions-policy/permissions-policy-header-policy-allowed-for-wildcard.https.sub.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src=/permissions-policy/resources/permissions-policy.js></script>
  <!-- Permissions-Policy: fullscreen=self wildcard_origin; -->
  <script>
  'use strict';
  var same_origin = 'https://{{domains[]}}:{{ports[https][0]}}';
  var wildcard_origin = 'https://*.{{domains[]}}:{{ports[https][0]}}';
  var cross_origin = 'https://{{domains[www]}}:{{ports[https][0]}}';
  var cross_origin1 = 'https://{{domains[www1]}}:{{ports[https][0]}}';
  var same_origin_src = '/permissions-policy/resources/permissions-policy-allowedfeatures.html';
  var cross_origin_src = cross_origin + same_origin_src;
  var cross_origin_src1 = cross_origin1 + same_origin_src;
  var header_policy = `Permissions-Policy: fullscreen=("${wildcard_origin}")`;

  // Test that fullscreen's allowlist is [self wildcard_origin].
  test(function() {
    assert_array_equals(
      document.featurePolicy.getAllowlistForFeature('fullscreen').sort(),
      [wildcard_origin, same_origin].sort());
  }, header_policy + ' -- test allowlist is [self wildcard_origin].');

  // Test that fullscreen is allowed on same-origin subframes with or without an allow attribute.
  test_allowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is allowed on same-origin subframe',
    'fullscreen',
    same_origin_src);
  test_allowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is allowed on same-origin subframe even with allow attribute',
    'fullscreen',
    same_origin_src,
    "fullscreen " + same_origin);

  // Test that fullscreen is disallowed on cross-origin subframes without an allow attribute.
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on cross-origin subframe',
    'fullscreen',
    cross_origin_src);
  test_allowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is allowed on cross-origin subframe allow attribute',
    'fullscreen',
    cross_origin_src,
    "fullscreen " + cross_origin);
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on another cross-origin subframe',
    'fullscreen',
    cross_origin_src1);
  test_allowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is allowed on another cross-origin subframe allow attribute',
    'fullscreen',
    cross_origin_src1,
    "fullscreen " + cross_origin1);

  // Test that wildcard allow attribute isn't supported.
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on cross-origin subframe with wildcard allow attribute',
    'fullscreen',
    cross_origin_src,
    "fullscreen " + wildcard_origin);
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on another cross-origin subframe with wildcard allow attribute',
    'fullscreen',
    cross_origin_src1,
    "fullscreen " + wildcard_origin);
  </script>
</body>
